requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
python-multipart==0.0.7
pydantic==2.4.2
pydantic-settings==2.0.3
torch==2.2.0
transformers==4.38.0
sentence-transformers==2.2.2
python-docx==1.0.1
PyMuPDF==1.23.5
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
elasticsearch==8.10.1
faiss-cpu==1.7.4
numpy==1.26.1
moviepy==1.0.3
python-pptx==0.6.21
SpeechRecognition==3.10.0
huggingface-hub==0.19.3
